Beautify Readme
When I see the First time, I see it was not proper readable, I had to give focus to read content categories and read the title, I have added thin lines between Every Heading or Title and change the Heading or Title tag to h3(###). so now it's is good to read and understandable also.